refactor: moved exported Nagware types to a new nagware-types file
